+++ src/main/java/io/supertokens/pluginInterface/passwordless/sqlStorage/PasswordlessSQLStorage.java
@@ -16,13 +16,16 @@
 
 package io.supertokens.pluginInterface.passwordless.sqlStorage;
 
+import io.supertokens.pluginInterface.bulkimport.exceptions.BulkImportTransactionRolledBackException;
 import io.supertokens.pluginInterface.emailpassword.exceptions.DuplicateEmailException;
 import io.supertokens.pluginInterface.emailpassword.exceptions.UnknownUserIdException;
 import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.pluginInterface.multitenancy.AppIdentifier;
 import io.supertokens.pluginInterface.multitenancy.TenantIdentifier;
+import io.supertokens.pluginInterface.multitenancy.exceptions.TenantOrAppNotFoundException;
 import io.supertokens.pluginInterface.passwordless.PasswordlessCode;
 import io.supertokens.pluginInterface.passwordless.PasswordlessDevice;
+import io.supertokens.pluginInterface.passwordless.PasswordlessImportUser;
 import io.supertokens.pluginInterface.passwordless.PasswordlessStorage;
 import io.supertokens.pluginInterface.passwordless.exception.DuplicatePhoneNumberException;
 import io.supertokens.pluginInterface.sqlStorage.SQLStorage;
@@ -30,6 +33,7 @@ import io.supertokens.pluginInterface.sqlStorage.TransactionConnection;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import java.util.List;
 
 public interface PasswordlessSQLStorage extends PasswordlessStorage, SQLStorage {
     PasswordlessDevice getDevice_Transaction(TenantIdentifier tenantIdentifier, TransactionConnection con,
@@ -85,4 +89,7 @@ public interface PasswordlessSQLStorage extends PasswordlessStorage, SQLStorage 
     void deletePasswordlessUser_Transaction(TransactionConnection con, AppIdentifier appIdentifier, String userId,
                                             boolean deleteUserIdMappingToo)
             throws StorageQueryException;
+
+    void importPasswordlessUsers_Transaction(TransactionConnection con, List<PasswordlessImportUser> users)
+            throws StorageQueryException, TenantOrAppNotFoundException, BulkImportTransactionRolledBackException;
 }