+++ CHANGELOG.md
@@ -7,6 +7,16 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [6.4.0] 
+
+- Adds support for Bulk Import
+- Adds `BulkImportUser` class to represent a bulk import user
+- Adds `BulkImportStorage` interface
+- Adds `DuplicateUserIdException` class
+- Adds `createBulkImportProxyStorageInstance` method in `Storage` class
+- Adds `closeConnectionForBulkImportProxyStorage`, `commitTransactionForBulkImportProxyStorage`, and `rollbackTransactionForBulkImportProxyStorage` method in `SQLStorage` class
+- Adds `BulkImportTransactionRolledBackException` for signaling if the transaction was rolled back by the DBMS
+
 ## [6.3.0] - 2024-10-02
 
 - Adds `OAuthStorage` interface for OAuth Provider support